Restructuring proxy so hooks are more orthogonal
This will make adding the CreateContainer hook cleaner.
